backend/requirements.txt
fastapi
uvicorn
pydantic
sse-starlette
google-genai
python-dotenv
